Experimental Sound Game in Java/Processing
"The Amazing Robo Dog at the Totally Inappropriate Cat Dance Party" is
an experimental sound game by Adam Rafinski, developed in
Java/Processing in Spring 2013 utilizing dynamic sound creation and AI
behaviors.

Current Version tested on Processing 2.2.1: https://www.processing.org

The player moves with "w, a, s, d" Robo Dog and can use the curser in
order to create speaker with sounds. The x coordinates of the mouse
determine the pitch of the tone, the y coordinates determines the size
and reach of the speaker. Then Cats appear to the party, each born with
an own "taste": They can either love the music and then they start to
dance - shake and become green - and beginn eventually to popular, or
they can be neutral about the music, or start to hate it, in which case
they turn red and start to destroy the speakers. After some experience
of anger the Cats may also start to attack Robo Dog. If Robo Dog is
touched, he loses a life. The player also has the ability to "distract"
attacking Cats with his mouse curser. The game ends when Robo Dog
succeeds in providing an amazing dance party in which Cats populate
like crazy and the game crashes. Since the crash is always determined
not only by the player's and the AI's behavior, but also by the machine
on which the game is running, the final moment appears always very
unique.